refactor(cli): split check command modules
